fix(blocks-construct): show extension banner in login page only
