+++ src/layout/auth-layout/auth-layout.tsx
@@ -1,20 +1,18 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { Outlet, useNavigate } from 'react-router-dom';
-import { AlertTriangle, SquareArrowOutUpRight, X } from 'lucide-react';
+import { AlertTriangle } from 'lucide-react';
 import bgAuthLight from '@/assets/images/bg_auth_light.svg';
 import bgAuthDark from '@/assets/images/bg_auth_dark.svg';
 import { useGetLoginOptions } from '@/modules/auth/hooks/use-auth';
 import { useAuthState } from '@/state/client-middleware';
 import { useTheme } from '@/styles/theme/theme-provider';
-import { Button } from '@/components/ui-kit/button';
-import { LanguageSelector } from '@/components/core';
+import { ExtensionBanner, LanguageSelector } from '@/components/core';
 
 export const AuthLayout = () => {
   const { isLoading, error: loginOptionsError } = useGetLoginOptions();
   const navigate = useNavigate();
   const { isMounted, isAuthenticated } = useAuthState();
   const { theme } = useTheme();
-  const [showBanner, setShowBanner] = useState(true);
 
   useEffect(() => {
     // Don't redirect if we're on the MFA verification page
@@ -144,51 +142,10 @@ export const AuthLayout = () => {
   };
 
   if (isLoading) return null;
-  const isBannerAllowedToVisible = [
-    'https://construct.seliseblocks.com/',
-    'https://stg-construct.seliseblocks.com',
-    'https://dev-construct.seliseblocks.com',
-  ].some((path) => window.location.href.startsWith(path));
 
   return (
     <div className="flex w-full flex-col h-screen">
-      {isBannerAllowedToVisible && showBanner && (
-        <div className="sm:relative w-full flex items-center sm:justify-between bg-surface py-3 px-4">
-          <div className="flex w-full items-center justify-center gap-2">
-            <span className="text-sm">
-              Experience UILM — explore its capabilities in Blocks Construct using the extension.{' '}
-              <a
-                href="https://selisegroup.com/blocks/"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="text-sm font-semibold underline hover:text-primary"
-              >
-                Learn more
-              </a>
-            </span>
-            <a
-              href="https://chromewebstore.google.com/detail/ehnhmdghlkaeaiinoahgipdeogkikjem?utm_source=item-share-cb"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              <Button variant="outline" size="sm" className="!gap-0">
-                <span className="capitalize font-bold text-sm hidden sm:inline">Get Extension</span>
-                <SquareArrowOutUpRight className="h-4 w-4 sm:ml-2" />
-              </Button>
-            </a>
-          </div>
-          <div className="sm:absolute sm:top-4 sm:right-4">
-            <Button
-              variant="ghost"
-              size="icon"
-              onClick={() => setShowBanner(false)}
-              className="h-fit w-fit p-1 rounded hover:bg-neutral-100 text-medium-emphasis"
-            >
-              <X className="h-4 w-4" />
-            </Button>
-          </div>
-        </div>
-      )}
+      <ExtensionBanner />
       <div className="flex w-full min-h-screen relative">
         <div className="hidden md:block w-[36%] relative bg-primary-50">
           <img
